refactor(rr): Factor common build pipeline out
